Working Swift 2.0 Model from Swagger. Huzzah. This commit adds the first working version of the automated Swift 2.0 model and API. There's still a lot to be done here, but this is a good first step.
The code compiles and it can be used to query the mail server against
the current list of mails. Other calls (for tags, or for drafts) are
also included but I haven't tested them yet.

The templates that generated this code are mustache templates and can
be found in

`swagger-codegen/modules/swagger-codegen/src/main/resources/swift`

The current template Swift code is sub par. We should work on improving
these templates for our own requirements such as Argo or BrightFutures or
whatever else imagine useful.